LayoutTests/fast/dom/Element/getBoundingClientRect-expected.txt

Client bounding rect for #1
testRect.top 42
testRect.left 8
testRect.width 300
testRect.height 100
testRect.bottom 142
testRect.right 308

PASS testRect.top is 42
PASS testRect.left is 8
PASS testRect.width is 300
PASS testRect.height is 100
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #2
testRect.top 192
testRect.left 8
testRect.width 320
testRect.height 120
testRect.bottom 312
testRect.right 328

PASS testRect.top is 192
PASS testRect.left is 8
PASS testRect.width is 320
PASS testRect.height is 120
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #3
testRect.top 362
testRect.left 18
testRect.width 300
testRect.height 100
testRect.bottom 462
testRect.right 318

PASS testRect.top is 362
PASS testRect.left is 18
PASS testRect.width is 300
PASS testRect.height is 100
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #4
testRect.top 491
testRect.left 37
testRect.width 142
testRect.height 142
testRect.bottom 633
testRect.right 179

PASS testRect.top is 491
PASS testRect.left is 37
PASS testRect.width is 142
PASS testRect.height is 142
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #5
testRect.top 662
testRect.left 8
testRect.width 300
testRect.height 240
testRect.bottom 902
testRect.right 308

PASS testRect.top is 662
PASS testRect.left is 8
PASS testRect.width is 300
PASS testRect.height is 240
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #6
Known failure. Skipping.

Client bounding rect for #7
testRect.top 1236
testRect.left 8
testRect.width 299
testRect.height 180
testRect.bottom 1416
testRect.right 307

PASS testRect.top is 1236
PASS testRect.left is 8
PASS testRect.width is 299
PASS testRect.height is 180
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #8
testRect.top 1466
testRect.left 8
testRect.width 300
testRect.height 46
testRect.bottom 1512
testRect.right 308

PASS testRect.top is 1466
PASS testRect.left is 8
PASS testRect.width is 300
PASS testRect.height is 46
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #9
testRect.top 1562
testRect.left 8
testRect.width 300
testRect.height 64
testRect.bottom 1626
testRect.right 308

PASS testRect.top is 1562
PASS testRect.left is 8
PASS testRect.width is 300
PASS testRect.height is 64
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #10
testRect.top 1678
testRect.left 8
testRect.width 300
testRect.height 20
testRect.bottom 1698
testRect.right 308

PASS testRect.top is 1678
PASS testRect.left is 8
PASS testRect.width is 300
PASS testRect.height is 20
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #11
testRect.top 1774
testRect.left 10
testRect.width 147
testRect.height 20
testRect.bottom 1794
testRect.right 157

PASS testRect.top is 1774
PASS testRect.left is 10
PASS testRect.width is 147
PASS testRect.height is 20
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #12
testRect.top 1868
testRect.left 8
testRect.width 100
testRect.height 100
testRect.bottom 1968
testRect.right 108

PASS testRect.top is 1868
PASS testRect.left is 8
PASS testRect.width is 100
PASS testRect.height is 100
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #13
testRect.top 2104
testRect.left 8
testRect.width 180
testRect.height 18
testRect.bottom 2122
testRect.right 188

PASS testRect.top is 2104
PASS testRect.left is 8
PASS testRect.width is 180
PASS testRect.height is 18
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #14
testRect.top 2172
testRect.left 8
testRect.width 300
testRect.height 236
testRect.bottom 2408
testRect.right 308

PASS testRect.top is 2172
PASS testRect.left is 8
FAIL testRect.width should be 200. Was 300.
PASS testRect.height is 236
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #15
testRect.top 2458
testRect.left 8
testRect.width 0
testRect.height 0
testRect.bottom 2458
testRect.right 8

PASS testRect.top is 2458
PASS testRect.left is 8
PASS testRect.width is 0
PASS testRect.height is 0
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

PASS successfullyParsed is true

TEST COMPLETE
